Move well name and symbol to util function instead of aquifer
